Release v0.3.8 - Documentation quality, fitting module refactor, CLI output fixes
- Bump version to 0.3.8
- Rewrite fitting.py docstrings to journal-article quality (no hedge stack, precise units)
- Refactor cli.py: convert user-facing logger.info to print() for capsys compatibility
- Refactor eprplot.py with improved docstrings and code structure
- Fix baseline/models.py: np.zeros_like dtype=float to handle integer input arrays (NumPy compat)
- Fix E301 blank-line errors in physics/constants.py, physics/units.py, signalprocessing/apowin.py, signalprocessing/frequency_analysis.py
- Fix E501 line-too-long in cli.py
- Update docs/conf.py, docs/index.rst, README.md for v0.3.8
- Update physics/constants.py and physics/units.py docstrings (CODATA 2022)
- Update isotope_gui.py and fair/conversion.py minor improvements

+++ epyr/baseline/models.py
@@ -115,7 +115,7 @@ def polynomial_1d(x, *coeffs):
     Returns:
         np.ndarray: Polynomial values
     """
-    result = np.zeros_like(x)
+    result = np.zeros_like(x, dtype=float)
     for i, coeff in enumerate(coeffs):
         result += coeff * (x**i)
     return result
